Class level configuration for Raix
This adds a class level configuration for Raix, allowing users to e.g. set the
client per class. This is useful if you want to use different API keys for
different parts of your application.

The code is a bit verbose and the `Raix::Configuration` class doesn't look that
nice anymore, but I felt that this would be the better choice over doing some
meta programming to first check the class level config and the fall back to the
global config. Obviously that is a debatable.

Not sure if everything is in the right place either and how much to test this.