Add ostruct dependency to prevent warning
In Ruby 3.4.4, this warning is emitted:

/Users/andy/.local/share/mise/installs/ruby/3.4.4/lib/ruby/3.4.0/ostruct.rb
was loaded from the standard library, but will no longer be part of the
default gems starting from Ruby 3.5.0.
You can add ostruct to your Gemfile or gemspec to silence this warning.
Also please contact the author of raix-0.9.1 to request adding ostruct
into its gemspec.